Use eventlet green threads for concurrency.
Adds support for eventlet green threads to improve concurrency in
small object workloads.

+++ s3_sync/s3_sync.py
@@ -1,3 +1,6 @@
+import eventlet
+eventlet.patcher.monkey_patch(all=True)
+
 import logging
 import os.path
 import sys
@@ -16,7 +19,7 @@ from .sync_container import SyncContainer
 class S3Sync(object):
     def __init__(self, conf):
         self.logger = logging.getLogger('s3-sync')
-        self.conf = conf
+        self.containers = conf.get('containers', [])
         self.root = conf['devices']
         self.interval = 10
         self.swift_dir = '/etc/swift'
@@ -27,7 +30,19 @@ class S3Sync(object):
         self.myips = whataremyips('0.0.0.0')
         self.items_chunk = conf['items_chunk']
         self.poll_interval = conf.get('poll_interval', 5)
+        self._setup_workers(conf)
         self.logger.debug('Created the S3Sync instance')
+
+    def _setup_workers(self, conf):
+        self.workers = conf.get('workers', 10)
+        self.pool = eventlet.GreenPool(self.workers)
+        self.work_queue = eventlet.queue.Queue(self.workers * 2)
+
+        # max_size=None means a Queue is infinite
+        self.error_queue = eventlet.queue.Queue(maxsize=None)
+        self.stats_queue = eventlet.queue.Queue(maxsize=None)
+        for _ in range(0, self.workers):
+            self.pool.spawn_n(self.worker)
 
     def get_broker(self, account, container, part, node):
         db_hash = hash_path(account, container)
@@ -36,48 +51,64 @@ class S3Sync(object):
                                db_hash + '.db')
         return ContainerBroker(db_path, account=account, container=container)
 
-    def sync_row(self, sync_container, row):
-        if row['deleted']:
-            return sync_container.delete_object(row['name'])
-        return sync_container.upload_object(row['name'],
-                                            row['storage_policy_index'])
+    def worker(self):
+        while 1:
+            work = self.work_queue.get()
+            if not work:
+                self.work_queue.task_done()
+                break
+            row, sync_container = work
+            try:
+                if row['deleted']:
+                    sync_container.delete_object(row['name'])
+                sync_container.upload_object(row['name'],
+                                             row['storage_policy_index'])
+            except Exception as e:
+                self.error_queue.put((row, e))
+            self.work_queue.task_done()
 
-    # TODO: use green threads here: need to understand whether it makes sense
+    def check_errors(self, account, container):
+        if self.error_queue.empty():
+            return
+        while not self.error_queue.empty():
+            row, error = self.error_queue.get()
+            if self.logger:
+                self.logger.error('Failed to propagate object %s/%s/%s: %s' % (
+                    account, container, row['name'], repr(error)))
+        raise RuntimeError('Failed to sync %s/%s' % (account, container))
+
     def sync_items(self, sync_container, rows, nodes_count, node_id):
-        errors = False
+        object_count = 0
+        start = time.time()
         for row in rows:
             if (row['ROWID'] % nodes_count) != node_id:
                 continue
+            object_count += 1
+            self.work_queue.put((row, sync_container))
             if self.logger:
                 self.logger.debug('propagating %s' % row['ROWID'])
-            try:
-                self.sync_row(sync_container, row)
-            except Exception as e:
-                self.logger.error('Failed to propagate row %s: %s' % (
-                    row['ROWID'], repr(e)))
-                errors = True
-        if errors:
-            raise RuntimeError('Failed to sync %s/%s' % (
-                sync_container.account, sync_container.container))
+        self.work_queue.join()
+        self.logger.info("Processed %d objects in %f seconds; %d errors" %
+                         (object_count, time.time() - start,
+                          self.error_queue.qsize()))
+        self.check_errors(sync_container.account, sync_container.container)
 
+        # TODO: Duplicated code that should be refactored.
+        object_count = 0
+        start = time.time()
         for row in rows:
             # Validate that changes from all other rows have also been sync'd.
             if (row['ROWID'] % nodes_count) == node_id:
                 continue
+            object_count += 1
+            self.work_queue.put((row, sync_container))
             if self.logger:
                 self.logger.debug('verifiying %s' % row['ROWID'])
-            try:
-                self.sync_row(sync_container, row)
-            except Exception as e:
-                self.logger.error('Failed to verify row %s: %s' %(
-                    row['ROWID'], repr(e)))
-                errors = True
-        if errors:
-            raise RuntimeError('Failed to verify %s/%s' % (
-                sync_container.account, sync_container.container))
-
-    def get_items_since(self, broker, since_start):
-        return broker.get_items_since(since_start, 10)
+        self.work_queue.join()
+        self.logger.info("Verified %d objects in %f seconds; %d errors" %
+                         (object_count, time.time() - start,
+                          self.error_queue.qsize()))
+        self.check_errors(sync_container.account, sync_container.container)
 
     def sync_container(self, container):
         part, container_nodes = self.container_ring.get_nodes(
@@ -100,6 +131,11 @@ class S3Sync(object):
                 container.save_status(items[-1]['ROWID'], broker_info['id'])
             return
 
+    def stop_pool(self):
+        for _ in range(0, self.workers):
+            self.work_queue.put(None)
+        self.pool.waitall()
+
     def run_always(self):
         self.logger.debug('Entering the poll loop')
         while True:
@@ -112,16 +148,18 @@ class S3Sync(object):
     def run_once(self):
         # Since we don't support reloading, the daemon should quit if there are
         # no containers configured
-        if 'containers' not in self.conf or not self.conf['containers']:
+        if not self.containers:
             sys.exit(0)
-        for sync_settings in self.conf['containers']:
+        for sync_settings in self.containers:
             try:
+                self.logger.debug('Sync container: %s/%s' % (
+                    sync_settings['account'], sync_settings['container']))
                 self.sync_container(SyncContainer(self.status_dir,
-                                                  sync_settings))
+                                                  sync_settings, self.workers))
             except Exception as e:
                 account = sync_settings.get('account', 'N/A')
                 container = sync_settings.get('container', 'N/A')
                 bucket = sync_settings.get('aws_bucket', 'N/A')
-                self.logger.error("Failed to sync %s/%s to %s: %s" %
-                    (account, container, bucket, repr(e)))
+                self.logger.error("Failed to sync %s/%s to %s: %s" % (
+                    account, container, bucket, repr(e)))
                 self.logger.error(traceback.format_exc(e))
